More yaks to shave
Makefile - Fixed a stupid typo in HTKMLFREADER name. Added the -Wno-dev
argument so we don't get yelled at about not setting up @rpath correctly.
A choice I'm sure that will come to haunt us.

SequenceParser.cpp, UCIParser.cpp, SerializationTests.cpp - Mac gets
picky about translating between long long unsigned int and size_t (even
though I think they are the same). So I need a cast.

+++ Source/Readers/LMSequenceReader/SequenceParser.cpp
@@ -404,7 +404,7 @@ size_t SequenceParser<NumType, LabelType>::UpdateBuffer()
     }
 
     // read the next block
-    size_t bytesToRead = min(m_bufferSize, m_fileSize - m_bufferStart) - saveBytes;
+    size_t bytesToRead = min(m_bufferSize, (size_t)m_fileSize - m_bufferStart) - saveBytes;
     size_t bytesRead = fread(m_fileBuffer + saveBytes, 1, bytesToRead, m_pFile);
     if (bytesRead == 0 && ferror(m_pFile))
         RuntimeError("SequenceParser::UpdateBuffer - error reading file");
